Migrate to Dart 2
This fixes type issues that prevented the server from running on Dart 2.

As part of this, I restructured `DeployConfig` to fail early if the
config is invalid and switched the GitHub webhook parsing to be based on
the `github` package we were already using for API requests.

I also added a set of common lints to enforce code style.